refactor: Double로 타입 변경

+++ OpenMarket/OpenMarket/Application/Domain/Model/Entities/ProductListEntity.swift
@@ -15,7 +15,7 @@ struct ProductEntity: Hashable {
     let thumbnailImage: UIImage
     let name: String
     let currency: String
-    let originalPrice: Int
-    let discountedPrice: Int
+    let originalPrice: Double
+    let discountedPrice: Double
     let stock: Int
 }
